registration

+++ myweb/base/views.py
@@ -1,10 +1,12 @@
-from django.shortcuts import render, redirect
-from .models import Room, Topic
+from django.shortcuts import render, redirect, HttpResponse
+from .models import Room, Topic, Messages
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 # Create your views here.
 def home(request):
@@ -18,10 +20,11 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=int(pk))
-    context = {'room': room}
+    room_messages = room.messages_set.all().order_by('-created')
+    context = {'room': room, 'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -34,9 +37,13 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("<h1>You don't have permission!</h1>")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -48,16 +55,22 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("<h1>You don't have permission!</h1>")
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html')
 
 def login_page(request):
+    page = "login"
     if request.method == "POST":
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -71,9 +84,36 @@ def login_page(request):
             return redirect('home')
         else:
             messages.error(request, "Password is uncorrect!")
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+
+def register_page(request):
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+    return render(request, 'base/login_register.html', {'form': form})
+
+
+
+@login_required(login_url='login')
+def delete_message(request, pk):
+    message = Messages.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse("<h1>You don't have permission!</h1>")
+
+    if request.method == "POST":
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
